requirements.txt
copasi_basico==0.78